requirements.txt
appdirs==1.4.3
backcall==0.1.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
cssselect==1.0.3
decorator==4.3.0
Django==2.1.3
django-environ==0.4.5
django-filter==2.0.0
django-money==0.14.3
djangorestframework==3.9.0
factory-boy==2.11.1
fake-useragent==0.1.11
Faker==1.0.0
flake8==3.6.0
idna==2.7
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.1
lxml==4.2.5
mccabe==0.6.1
parse==1.9.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
py-moneyed==0.8.0
pycodestyle==2.4.0
pyee==5.0.0
pyflakes==2.0.0
Pygments==2.3.0
pyppeteer==0.0.25
pyquery==1.4.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
six==1.11.0
text-unidecode==1.2
tqdm==4.28.1
traitlets==4.3.2
urllib3==1.24.1
w3lib==1.19.0
wcwidth==0.1.7
websockets==7.0
